\name{RcmdrEcoVirtualPlugin-internal}
\title{Internal RcmdrEcoVirtualPlugin objects}
\alias{estDemDb}
\alias{popExpDb}
\alias{popLogDb}
\alias{logBifDb}
\alias{popStrDb}
\alias{propRain}
\alias{intCol}
\alias{resEff}
\alias{resEffcol}
\alias{spDep}
\alias{compDb}
\alias{metacompDb}
\alias{comcompDbox}
\alias{regnichoDbox}
\alias{sucmatrixDbox}
\alias{ColExtDbox}
\alias{archipDbox}
\alias{bioGeoIslDbox}
\alias{extGameDbox}
\alias{gr.abund}
\alias{hubDbox}
\alias{hubDbox2}
\alias{hubDbox3}
\alias{randWalkDbox}
\alias{estEnvDb}

\description{Internal RcmdrEcoVirtualPlugin objects.}
\details{These are not to be called by the user. See the help on packages Rcmdr and EcoVirtual for details.}

\author{
Alexandre Adalardo de Oliveira \email{ecovirtualpackage@gmail.com}
}
\examples{
\dontrun{
Commander()
}
}


\seealso{
        \url{http://ecovirtual.ib.usp.br}
}
\keyword{internal}
